Removing duplicated code
