+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/list/background/ShapeAppearanceModelFactory.kt
@@ -17,7 +17,12 @@
 package io.getstream.chat.android.ui.feature.messages.list.background
 
 import android.content.Context
+import androidx.core.content.ContextCompat
+import androidx.core.graphics.drawable.DrawableCompat.setTint
+import com.google.android.material.shape.MaterialShapeDrawable
 import com.google.android.material.shape.ShapeAppearanceModel
+import io.getstream.chat.android.ui.R
+import io.getstream.chat.android.ui.utils.extensions.dpToPxPrecise
 import io.getstream.chat.android.ui.utils.extensions.isRtlLayout
 
 /**
@@ -60,4 +65,25 @@ internal object ShapeAppearanceModelFactory {
             }
             .build()
     }
+
+    fun fileBackground(context: Context): MaterialShapeDrawable = roundedBackground(context)
+
+    fun audioBackground(context: Context): MaterialShapeDrawable = roundedBackground(context)
+
+    private fun roundedBackground(context: Context): MaterialShapeDrawable =
+        ShapeAppearanceModel.builder()
+            .setAllCornerSizes(CORNER_SIZE_PX)
+            .build()
+            .let(::MaterialShapeDrawable)
+            .apply {
+                setStroke(
+                    STROKE_WIDTH_PX,
+                    ContextCompat.getColor(context, R.color.stream_ui_grey_whisper)
+                )
+                setTint(ContextCompat.getColor(context, R.color.stream_ui_white))
+            }
+
 }
+
+private val CORNER_SIZE_PX = 12.dpToPxPrecise()
+private val STROKE_WIDTH_PX = 1.dpToPxPrecise()
